guides/resources/module/connoisseur/connoisseur/requirements.txt
#connoisseur requirments
#Cython==0.24.1
numpy==1.12.0
scipy==0.18.1
scikit-image==0.12.3
matplotlib==1.5.3
#ipython==3.0.0
h5py==2.6.0
leveldb==0.20
lmdb==0.92
networkx==1.7
#nose==1.3.0
pandas==0.19.2
python-dateutil==1.5
protobuf==2.5.0
python-gflags==2.0
pyyaml==3.10
pyvoro==1.3.2
Pillow==4.0.0
six==1.10
dask[array]==0.15
cycler==0.10.0
toolz==0.8.2
shapely==1.6.4.post1
#caffe # there's really no pypy package for caffe
